Make GCC (-O2) happy

+++ src-api/common/isonumber.c
@@ -51,7 +51,7 @@
 #include "common/string.h"
 
 static const char *_isonumber_u64_to_string(char *out,
-    size_t out_len, uint64_t number, const char *unit, int fraction,
+    size_t out_len, uint64_t number, const char *unit, size_t fraction,
     bool binary, bool raw);
 
 /**
@@ -266,7 +266,7 @@ isonumber_to_u64(uint64_t *dst, const char *iso, int fraction, bool binary) {
  */
 static const char *
 _isonumber_u64_to_string(char *out, size_t out_len,
-    uint64_t number, const char *unit, int fraction,
+    uint64_t number, const char *unit, size_t fraction,
     bool binary, bool raw) {
   static const char symbol_large[] = " kMGTPE";
   static const char symbol_small[] = " munpfa";
